🐛 FIX:  signup handler and logics
