+++ client/pages/signup.tsx
@@ -4,6 +4,7 @@ import router from 'next/router';
 import useForm from '../hooks/useForm';
 import { useMutation } from '@apollo/client';
 import { SIGNUP } from '../utils/queries/signup';
+// import { useStateValue } from '../data/State.context';
 
 const Signup: FC = () => {
   const [username, handleUsername] = useForm('');
@@ -12,6 +13,7 @@ const Signup: FC = () => {
   const [avatar, handleAvatar] = useForm('');
   const [remember, setRemember] = useState(false);
   const [signUp] = useMutation(SIGNUP);
+  // const [, disptach] = useStateValue();
 
   const createKarbanAccount = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
@@ -22,8 +24,10 @@ const Signup: FC = () => {
 
       if (errors) return; // !FIX: better error handlings
 
-      localStorage.setItem('KARBAN_TOKEN', data.token);
-      router.push('/dashboard');
+      localStorage.setItem('KARBAN_TOKEN', data.signUp.token);
+
+      // disptach({ type: 'SET_USER', payload: data });
+      // router.push('/dashboard');
     }
   };
